+++ packages/guide/src/app/database/database.module.ts
@@ -25,6 +25,7 @@ import { LocationsModule } from '../locations/locations.module';
 import { DatabaseShopDataComponent } from './components/database-shop-data/database-shop-data.component';
 import { DatabaseShopItemProcessComponent } from './components/database-shop-item-process/database-shop-item-process.component';
 import { DatabaseItemUpgradeComponent } from './components/database-item-upgrade/database-item-upgrade.component';
+import { DatabaseBestiaryComponent } from './components/database-bestiary/database-bestiary.component';
 
 @NgModule({
     declarations: [
@@ -44,6 +45,7 @@ import { DatabaseItemUpgradeComponent } from './components/database-item-upgrade
         DatabaseShopDataComponent,
         DatabaseShopItemProcessComponent,
         DatabaseItemUpgradeComponent,
+        DatabaseBestiaryComponent,
     ],
     imports: [
         CommonModule,